requirements.txt
numpy
keras
opencv-python
pillow
Flask
rasa